Avoid having default PersonTypes being modified/deleted

+++ src/main/java/be/janschraepen/hellokitty/services/impl/PersonTypeServiceImpl.java
@@ -1,5 +1,6 @@
 package be.janschraepen.hellokitty.services.impl;
 
+import be.janschraepen.hellokitty.domain.persontype.CannotModifyPersonTypeException;
 import be.janschraepen.hellokitty.domain.persontype.ObjectFactory;
 import be.janschraepen.hellokitty.domain.persontype.PersonType;
 import be.janschraepen.hellokitty.domain.persontype.PersonTypeDTO;
@@ -45,14 +46,20 @@ public class PersonTypeServiceImpl implements PersonTypeService {
     }
 
     @Override
-    public PersonTypeDTO savePersonType(PersonTypeDTO dto) {
+    public PersonTypeDTO savePersonType(PersonTypeDTO dto) throws CannotModifyPersonTypeException {
         PersonType personType;
         if (StringUtils.isBlank(dto.getId())) {
             // save a new personType
             personType = new PersonType();
+            if (isSystemPersonType(dto.getShortCode())) {
+                throw new CannotModifyPersonTypeException();
+            }
         } else {
             // update an existing personType
             personType = personTypeRepository.findById(dto.getId());
+            if (isSystemPersonType(personType.getShortCode())) {
+                throw new CannotModifyPersonTypeException();
+            }
         }
         personType.setShortCode(dto.getShortCode());
         personType.setName(dto.getName());
@@ -62,18 +69,32 @@ public class PersonTypeServiceImpl implements PersonTypeService {
     }
 
     @Override
-    public void deletePersonType(String uuid) {
+    public void deletePersonType(String uuid) throws CannotModifyPersonTypeException {
         PersonType personType = personTypeRepository.findById(uuid);
         if (personType != null) {
+            if (isSystemPersonType(personType.getShortCode())) {
+                throw new CannotModifyPersonTypeException();
+            }
             personTypeRepository.delete(personType);
         }
     }
 
     @Override
-    public void deletePersonTypes(String[] uuids) {
+    public void deletePersonTypes(String[] uuids) throws CannotModifyPersonTypeException {
         for (String uuid : uuids) {
             deletePersonType(uuid);
         }
+    }
+
+    /**
+     * Check wh
+     * @param shortCode
+     * @return
+     */
+    boolean isSystemPersonType(String shortCode) {
+        return PersonType.OWNER.equals(shortCode)
+                || PersonType.CONTACT.equals(shortCode)
+                || PersonType.VET.equals(shortCode);
     }
 
 }
